requirements.txt
fastapi
uvicorn
sqlalchemy
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
black
mypy
flake8
requests
pytest
pytest-cov
isort
autoflake
google-api-python-client
google-cloud-translate